+++ NEWS.md
@@ -1,4 +1,15 @@
+# shinyCox 1.1.0
+
+* Added support for confidence intervals in plots
+* Added ability to change y-axis label in shiny app
+* Added new functions to support the addition of confidence intervals
+* Changed license from MIT to LGPL-3.0
+* Added `library(survival)` and `library(shinyCox)` to shiny app output to fix
+  errors when packages were not loaded
+
 # shinyCox 1.0.1
+
+* Submitted to CRAN, working on updates
 
 # shinyCox 1.0.0